% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_formula.janus}
\alias{expand_formula.janus}
\title{Expand the formula associated with a janus fitted model object.}
\usage{
\method{expand_formula}{janus}(model)
}
\arguments{
\item{model}{A fitted model in an object of class janus.}
}
\value{
A formula object containing the expanded formula.
}
\description{
If a formula is passed to fit in a variable, the variable will be stored in
the resulting fitted model object's call variable. This method is a
convenient utility for extracting and expanding the formula from a fitted
janus model object.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

