% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Feature selection through filtering}
\usage{
filter(formula, data, method = c("pearson", "spearman", "chisq", "infogain",
  "gainratio", "cfs"), ..., limit = NULL)
}
\arguments{
\item{formula,}{a formula object}

\item{data,}{a data frame}

\item{method,}{a string indicating the filter method to use}

\item{...,}{arguments to be passed on to called functions}

\item{limit,}{an optional integer specifying the number of features to retain}
}
\description{
Feature selection through filtering
}

