% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_cfs.R
\name{filter_cfs}
\alias{filter_cfs}
\title{Feature selection using CFS algorithm.}
\usage{
filter_cfs(formula, data)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing predictor and response variables.}
}
\value{
A vector of predictor identifiers.
}
\description{
Uses the correlation-based feature selection algorithm as implemented in the
FSelector package to select important features.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

