% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_chisq.R
\name{filter_chisq}
\alias{filter_chisq}
\title{Univariate filter using chi-squared.}
\usage{
filter_chisq(formula, data, limit)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing predictor and response variables.}

\item{limit}{Optional integer specifying the number of features to retain.}
}
\value{
A vector of feature importance values.
}
\description{
Uses chi-square for feature selection, as implemented in the FSelector
package, to select important features.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

