% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_spearman.R
\name{filter_spearman}
\alias{filter_spearman}
\title{Univariate filter using Spearman rank correlation.}
\usage{
filter_spearman(formula, data, limit)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing predictor and response variables.}

\item{limit}{Optional integer specifying the number of features to retain.}
}
\value{
A vector of feature importance values.
}
\description{
Uses Spearman's rank correlation for feature selection, as implemented in the
FSelector package, to select important features.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

