% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit.default}
\alias{fit.default}
\title{Fit a classification model using separate predictor and
   response variables.}
\usage{
\method{fit}{default}(x, y, classifier = c("glmnet"), subset, ...)
}
\arguments{
\item{x}{A dataframe or matrix containing the predictor variables.}

\item{y}{A vector or matrix of output variables.}

\item{classifier}{A string indicating the classifier to use. Available
   options (and source package) include:
\enumerate{
  \item "glm": logistic regression (glm)
  \item "glmnet": binomial or multinomial logistic regression (glmnet)
  \item "randomForest": random forest (randomForest)
  \item "e1071": support vector machine (e1071)
}}

\item{subset}{Vector of indices for the subset to be used in fitting.}

\item{...}{Additional arguments to be passed to classifiers.}
}
\value{
A fitted model in an object of class janus.
}
\description{
Uses a dataframe or matrix of predictor features to train a model using
a vector (for binomial) or matrix or dataframe (for multinomial) regression.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

