% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit.formula}
\alias{fit.formula}
\title{Fit a classification model using a formula object.}
\usage{
\method{fit}{formula}(formula, data, classifier = c("e1071", "glm",
  "randomforest", "glmnet"), subset, ...)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing the predictor and response variables.}

\item{classifier}{A string indicating the classifier to use. Available
   options (and source package) include:
\enumerate{
  \item "glm": logistic regression (glm)
  \item "randomForest": random forest (randomForest)
  \item "e1071": support vector machine (e1071)
}}

\item{subset}{Vector of indices for the subset to be used in fitting.}

\item{...}{Additional arguments to be passed to classifiers.}
}
\value{
A fitted model in an object of class janus.
}
\description{
Fit a classification model using a formula object.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

