% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit}
\alias{fit}
\title{Fit a classification model.}
\usage{
fit(x, ...)
}
\arguments{
\item{x}{Either a formula object or a database or matrix object.}

\item{...}{Additional parameters to be passed on.}
}
\value{
A fitted model in object of class janus.
}
\description{
Generic method for fitting a classification model.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

