% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_e1071.R
\name{fit_e1071}
\alias{fit_e1071}
\title{Fit a support vector machine classifier using e1071::svm.}
\usage{
fit_e1071(formula, data, ...)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing the predictor variables and a categorical
response variable.}

\item{...}{Arguments to pass to called functions.}
}
\value{
A fitted model in object of class janus.
}
\description{
Uses svm from the e1071 package to fit a support vector machine. The type
argument to svm is set to force classification. Also, the function uses
linear as the default to the kernel argument rather than radial as in native
svm. The cost parameter retains its default (i.e. cost = 1) and can be
modified by passing through a different value.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

