% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit a two-class logistic regression model using stats::glm}
\usage{
fit_glm(formula, data)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing the predictor variables and a categorial
output variable.}
}
\value{
A fitted model in object of class janus.
}
\description{
Fit a two-class logistic regression model using stats::glm
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

