% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_glmnet.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{Fit a binomial or multinomial logistic regression classifier using
   glmnet::glmnet.}
\usage{
fit_glmnet(x, y, ..., cvfit = TRUE, type_measure = "deviance", folds = 10)
}
\arguments{
\item{x}{A dataframe or matrix containing predictors.}

\item{y}{A vector containing response variables with two or more
levels.}

\item{...}{Additional parameters to pass to glmnet.}

\item{cvfit}{Logical indicating whether to use cross validation to find best
value for model hyperparameters. Defaults to TRUE.}

\item{type_measure}{String indicating the measure to use during cross
validation. Default measure is "deviance". Alternatives include:
\enumerate{
  \item "class": misclassification error rate
  \item "auc": area under the curve, for two-class models only
  \item "mse": mean squared error
  \item "mae": mean absolute error
}}

\item{folds}{The number of folds to use during cross validation. Default
   number of folds is 10.}
}
\value{
A fitted model in object of class janus.
}
\description{
Fit a binomial or multinomial logistic regression classifier using
   glmnet::glmnet.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

