% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_randomforest.R
\name{fit_randomforest}
\alias{fit_randomforest}
\title{Fit a support vector machine classifier using randomForest::randomForest}
\usage{
fit_randomforest(formula, data, ...)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A data frame containing the predictor variables and a categorical
response variable.}

\item{...}{Additional parameters to be passed on.}
}
\value{
A fitted model in object of class janus.
}
\description{
Fit a support vector machine classifier using randomForest::randomForest
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

