% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.janus}
\alias{plot.janus}
\title{Plot a janus object.}
\usage{
\method{plot}{janus}(x, ...)
}
\arguments{
\item{x}{A janus object containing a fitted model.}

\item{...}{Additional plotting parameters.}
}
\value{
Nothing
}
\description{
Plot a janus object.
}
\examples{
# create predictor data matrix and response data vector
x <- mtcars[, -9]
y <- mtcars[, 9]

# fit a regularlized cv.glmnet classifier, using cross-validation for
# determining the optimal value for model hyperparameters
mod <- fit(x, y, classifier = "glmnet")

# plot the model error measure against values for the penalty glmnet penalty
# paramter
plot(mod)
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

