% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_glmnet.R
\name{plot_glmnet}
\alias{plot_glmnet}
\title{Plot a janus object derived from a glmnet or cv.glmnet object}
\usage{
plot_glmnet(object, ...)
}
\arguments{
\item{object}{A janus object containing a fitted model.}

\item{...}{Additional plotting parameters.}
}
\value{
Nothing
}
\description{
Plot a janus object derived from a glmnet or cv.glmnet object
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

