% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.janus}
\alias{predict.janus}
\title{Predict probabilities or class labels for a fitted janus model}
\usage{
\method{predict}{janus}(object, newdata, type = c("class", "probability"))
}
\arguments{
\item{object}{a janus object containing a fitted model, as produced by fit()}

\item{newdata}{a dataframe that will be used to predict class labels}

\item{type}{a string indicating the desired output, either class labels or
   probabilities}
}
\value{
a data structure containining class labels or probabilities, as
   determined by the type argument
}
\description{
Uses a janus model object to predict class labels. If new data is supplied,
the class labels or probabilities for each option are returned. Otherwise
the label or probabilities for the training data are predicted.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

