% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_e1071.R
\name{predict_e1071}
\alias{predict_e1071}
\title{Predict response for svm model in janus object.}
\usage{
predict_e1071(object, newdata, type, ...)
}
\arguments{
\item{object}{A janus object containing a fitted glmnet or cv.glmnet model.}

\item{newdata}{A dataframe that will be used to predict either probabilities
for each class or class labels.}

\item{type}{A string indicating the desired output: either "class" for class
labels or "probabilities" for probabilities.}

\item{...}{Arguments to be passed on to underlying functions.}

\item{threshold}{Probability threshold for binary classification. Th default
  value is 0.5.}
}
\value{
A vector or dataframe containining class labels or probabilities, as
  determined by the type argument.
}
\description{
Makes predictions for models trained using the svm classifier in the package
e1071.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

