% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_glm.R
\name{predict_glm}
\alias{predict_glm}
\title{Predict response for glm model in janus object.}
\usage{
predict_glm(object, newdata, type, threshold)
}
\arguments{
\item{object}{A janus object containing a fitted glmnet or cv.glmnet model.}

\item{newdata}{A dataframe that will be used to predict either probabilities
for each class or class labels.}

\item{type}{A string indicating the desired output: either "class" for class
labels or "probabilities" for probabilities.}

\item{threshold}{Probability threshold for binary classification. Th default
  value is 0.5.}

\item{...}{Arguments to be passed on to underlying functions.}
}
\value{
A vector or dataframe containining class labels or probabilities, as
  determined by the type argument.
}
\description{
Makes predictions for models trained using the glm classifier in the package
stats.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

