% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_glmnet.R
\name{predict_glmnet}
\alias{predict_glmnet}
\title{Predict response for glmnet model in janus object.}
\usage{
predict_glmnet(object, newdata, type, ..., threshold)
}
\arguments{
\item{object}{A janus object containing a fitted glmnet or cv.glmnet model.}

\item{newdata}{A dataframe or matrix that will be used to predict either
probabilities for each class or class labels. If supplied as a dataframe,
it will be coerced to a matrix. This is required for obtained predictions
for either testing or training data.}

\item{type}{A string indicating the desired output: either "class" for class
labels or "probabilities" for probabilities.}

\item{...}{Arguments to be passed on to underlying functions.}

\item{threshold}{Probability threshold for binary classification. Th default
   value is 0.5.}
}
\value{
A vector or dataframe containining class labels or probabilities, as
   determined by the type argument.
}
\description{
Makes predictions for models trained using the glmnet logistic regression
classifier. This function handles both two-class and multinomial logistic
regression. By default, the predictions are made using the model with the
optimal regularization by setting the glmnet "s" parameter to the minimum
value of lambda obtained by cross validation using the cv.glmnet classifier.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

