% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.janus}
\alias{summary.janus}
\title{Summarize the output of a janus model}
\usage{
\method{summary}{janus}(object, ...)
}
\arguments{
\item{object}{An object of class janus}

\item{...}{Additional parameters to be passed on.}
}
\value{
None
}
\description{
Provides a text summary of a fitted janus model, with appropriate output
   based on the underlying classifier.
}
\examples{
# train a logistic regression model using glm
mod <- fit(am ~ mpg, data = mtcars, classifier = "glm")

# show a summary for the fitted model
summary(mod)
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

