% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_formula}
\alias{expand_formula}
\title{Generic method for expanding the formula associated with a model object.}
\usage{
expand_formula(model)
}
\arguments{
\item{model}{A fitted model constructed using a formula interface.}
}
\description{
Generic method for expanding the formula associated with a model object.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

