% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Feature selection through filtering}
\usage{
filter(formula, data, method = c("pearson", "spearman", "chisq", "infogain",
  "gainratio", "cfs"), ..., limit = NULL)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing predictor and response variables.}

\item{method}{A string indicating the filter method to use. One of:
\enumerate{
  \item "cfs": correlation-based feature selector (from FSelector package)
  \item "chisq": chi-square (FSelector)
  \item "gainratio": information gain ratio (FSelector)
  \item "infogain": information gain (FSelector)
  \item "pearson": Pearson correlation (FSelector)
  \item "spearman": Spearman rank-based correlation (FSelector)
}}

\item{...}{Arguments to be passed on to called functions.}

\item{limit}{Optional integer specifying the number of features to retain.}
}
\value{
A vector of feature importance values or predictor identifiers,
  depending on the chosen method.
}
\description{
Feature selection through filtering
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

