% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Feature selection through filtering.}
\usage{
filter(formula, data, method = c("pearson", "spearman", "chisq", "infogain",
  "gainratio", "cfs"), ..., limit = NULL)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe containing predictor and response variables.}

\item{method}{A string indicating the filter method to use. One of:
\itemize{
  \item "cfs": correlation-based feature selector (from FSelector package)
  \item "chisq": chi-square (FSelector)
  \item "gainratio": information gain ratio (FSelector)
  \item "infogain": information gain (FSelector)
  \item "pearson": Pearson correlation (FSelector)
  \item "spearman": Spearman rank-based correlation (FSelector)
}}

\item{...}{Arguments to be passed on to called functions.}

\item{limit}{Optional integer specifying the number of features to retain.}
}
\value{
A vector of feature importance values or predictor identifiers,
  depending on the chosen method.
}
\description{
Allows choosing from a range of algorithms for univariate feature selection.
Also allows use of the CFS algorithm for choosing the most important feature
in a set of predictor variables.
}
\examples{
# set up an example dataframe
set.seed(1)
y <- sample(0:1, 100, replace = TRUE)
x1 <- rnorm(100)
x2 <- sample(0:1, 100, replace = TRUE)
df <- data.frame(y, x1, x2)

# use Pearson's correlation for filtering
filter(y ~ ., df, method = "pearson")

# use Spearman's rank correlation and limit the number of returned variables
filter(y ~ ., df, method = "spearman", limit = 1)

# use the CFS algorithm to return the most important predictor
filter(y ~ ., df, method = "cfs")
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

