% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit}
\alias{fit}
\title{fit a model of the specified type
currently only supports passing a formula object and data in a data.frame}
\usage{
fit(formula, data, classifier = c("e1071", "glm", "randomforest"), subset,
  ...)
}
\arguments{
\item{formula,}{a formula object}

\item{data,}{a data frame}

\item{classifier,}{a string indicating the classifier to use}

\item{subset,}{vector of indices for the subset to be used in fitting}

\item{...,}{additional arguments to be passed to classifiers}
}
\value{
fitted model in object of class janus
}
\description{
fit a model of the specified type
currently only supports passing a formula object and data in a data.frame
}

