% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit}
\alias{fit}
\title{Fit a classification model.}
\usage{
fit(x, ...)
}
\arguments{
\item{x}{Either a formula object or a database or matrix object.}

\item{...}{Additional parameters to be passed on.}
}
\value{
A fitted model in object of class janus.
}
\description{
Generic method for fitting a classification model.
}
\examples{
# create a dataframe containing no missing values
mt_complete <- mtcars[complete.cases(mtcars), ]

# train a logistic regression model using glm
fit(am ~ mpg, data = mt_complete, classifier = "glm")

# train a random forest classifier using randomForest
fit(as.factor(am) ~ ., data = mt_complete, classifier = "randomforest")

# train a support vector machine with linear kernel using e1071::svm
fit(am ~ ., data = mt_complete, classifier = "e1071")
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}
\seealso{
Helper functions \code{\link{fit_e1071}}, \code{\link{fit_glm}},
  \code{\link{fit_glmnet}}, \code{\link{fit_randomforest}}
}

