% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.janus}
\alias{plot.janus}
\title{Plot a janus object.}
\usage{
\method{plot}{janus}(object)
}
\arguments{
\item{object}{A janus object containing a fitted model.}
}
\value{
Nothing
}
\description{
Plot a janus object.
}
\author{
Alex Wollenschlaeger, \email{alexw@panix.com}
}

